requirements.txt
python-telegram-bot==13.7
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
psutil==5.9.6
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.13.0